modify

+++ app/TX/interface.py
@@ -701,7 +701,7 @@ def _check_balance(address,assetId,value):
     if balances:
         for b in balances:
             if b.get("asset") == assetId:
-                balance = int(b.get("value"))
+                balance = float(b.get("value"))
     if balance < int(value):
         return False
     return True
